+++ examples/webpack-mapzen-worker/index.ts
@@ -1,16 +1,13 @@
 import "core-js/stable";
-import "cesiumSource/Widgets/widgets.css";
-import "./main.css";
 // @ts-ignore
-import * as Cesium from "cesiumSource/Cesium";
-// Import @types/cesium to use along with CesiumJS
-// @ts-ignore
-import { MartiniTerrainProvider } from "lib/terrain-provider";
-// @ts-ignore
-import { DefaultHeightmapResource } from "lib/resources/heightmap-resource";
-// @ts-ignore
-import { WorkerFarmTerrainDecoder } from "lib/worker/decoder";
+import {
+  MartiniTerrainProvider,
+  DefaultHeightmapResource,
+  WorkerFarmTerrainDecoder,
+} from "lib";
+
 import TerrariumWorker from "./mapzen.worker";
+import { buildExample } from "../_shared";
 
 // Mapzen API discontinued, alternate source required
 const terrainResource = new DefaultHeightmapResource({
@@ -34,61 +31,4 @@ const terrainProvider = new MartiniTerrainProvider({
   decoder: terrainDecoder,
 });
 
-let satellite = new Cesium.MapboxImageryProvider({
-  mapId: "mapbox.satellite",
-  maximumLevel: 19,
-  accessToken: process.env.MAPBOX_API_TOKEN,
-});
-
-const opts = {
-  terrainProvider, //: createWorldTerrain(),
-  // imageryProvider: Cesium.createWorldImagery({
-  //   style: Cesium.IonWorldImageryStyle.AERIAL,
-  // }),
-  // @ts-ignore
-  skyBox: false as false,
-  baseLayerPicker: false,
-  geocoder: false,
-  skyAtmosphere: false as false,
-  animation: false,
-  timeline: false,
-  // Makes cesium not render high fps all the time
-  requestRenderMode: true,
-  // Use full scene buffer (respecting pixel ratio) if this is false
-  useBrowserRecommendedResolution: false,
-  // We have a bug in the tile bounding box calculation somewhere.
-  terrainExaggeration: 1.0,
-  baseLayer: new Cesium.ImageryLayer(satellite),
-};
-
-const domID = "cesium-container";
-const g = document.createElement("div");
-g.id = domID;
-document.body.appendChild(g);
-
-const clat = -21.133786;
-const clon = 14.5481193;
-
-const viewer = new Cesium.Viewer(domID, opts);
-// Quadtree props: don't preload ancestors
-
-//viewer.scene.globe.baseColor = Cesium.Color.AQUAMARINE
-// @ts-ignore
-//viewer.scene.globe._surface._tileProvider._debug.wireframe = true
-// @ts-ignore
-viewer.extend(Cesium.viewerCesiumInspectorMixin);
-viewer.scene.debugShowFramesPerSecond = true;
-
-const extent = Cesium.Cartesian3.fromDegrees(clon, clat - 0.3, 8000);
-viewer.camera.setView({
-  destination: extent,
-  orientation: {
-    heading: Cesium.Math.toRadians(0), // east, default value is 0.0 (north)
-    pitch: Cesium.Math.toRadians(-15), // default value (looking down)
-    roll: 0.0, // default value
-  },
-});
-
-//viewer.resolutionScale = 2
-//viewer.scene.globe.enableLighting = true
-//viewer.canvas.style.imageRendering = false
+buildExample(terrainProvider, process.env.MAPBOX_API_TOKEN);
