+++ examples/webpack-react/src/index.ts
@@ -1,7 +1,7 @@
 import "core-js/stable";
 import "./main.css";
-import "cesiumSource/Widgets/widgets.css";
-const Cesium: any = require("cesiumSource/Cesium");
+import "cesium/Source/Widgets/widgets.css";
+const Cesium: any = require("cesium/Source/Cesium");
 // Import @types/cesium to use along with CesiumJS
 import { MapboxImageryProvider } from "cesium";
 import TerrainProvider from "../../../dist";
