--- examples/webpack/src/main.css
@@ -1,11 +1,0 @@
-body,
-#cesium-container {
-  width: 100vw;
-  height: 100vh;
-  margin: 0;
-  padding: 0;
-}
-
-.cesium-cesiumInspector {
-  transform: translate(0, 70px);
-}
