+++ src/index.ts
@@ -6,6 +6,8 @@ import {
   StretchedTilingScheme,
 } from "./terrain-provider";
 import MapboxTerrainProvider from "./mapbox-terrain-provider";
+export * from "./worker/decoder";
+export * from "./worker/worker-util";
 
 export default MapboxTerrainProvider;
 export {
